backend/requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
pydantic-settings>=2.11.0
python-multipart>=0.0.20
groq>=0.31.0
pymupdf>=1.26.4
python-docx>=1.2.0
python-pptx>=1.0.2
spacy>=3.8.7
language-tool-python>=2.9.4
langchain>=0.3.27
langgraph>=0.6.7
pytest>=8.4.2
httpx>=0.28.1
